Create abstarct View
Create a new class ViewTerminal and move there all code from View.
View transform to abstract class as a parent
of class ViewTerminal.
